LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/invalid-header.https-expected.txt



FAIL Verify the response of FetchEvent using XMLHttpRequest assert_equals: expected "finish" but got "failure:XHR must fail."
